feat(js): wasm64 (memory64) build pipeline — Gemma 3 4B runs
The wasm32 build can't load models whose total working set exceeds
4 GiB (model tensors + mmproj + KV cache + compute > 4 GiB). Gemma
3 4B Q4_K_M + its mmproj is ~5 GiB; on wasm32 it panics with
`raw_vec::capacity_overflow` during llama.cpp's tensor load. This
adds a sibling wasm64 build pipeline that ships a memory64 wasm
artifact capable of running such models — verified end-to-end with
Gemma 3 4B on Node.

What's in
---------

Sibling build script:
  - `nobodywho/js/scripts/build-pkg-emscripten-wasm64.sh`
  - `nobodywho/js/targets/wasm64-unknown-emscripten.json` (custom
    Rust target spec, tier-3 via `-Z build-std`)
  - `nobodywho/js/scripts/modelpath-smoke-wasm64.mjs` (smoke against
    the produced artifacts)
  - Output dir `nobodywho/js/pkg-bundler-wasm64/` (gitignored,
    parallel to existing `pkg-bundler/`)

Workspace cfg-gate refactor (backwards-compatible):
  - `cfg(target_arch = "wasm32")` → `cfg(target_family = "wasm")`
    across ~30 sites in core/ and js/. The new form matches BOTH
    wasm32 and wasm64; existing wasm32 build keeps the same
    behavior. Files touched: `core/Cargo.toml`, `core/src/{chat,
    crossencoder,encoder,llm,memory,tokenizer,tool_calling/mod}.rs`,
    `js/src/lib.rs`, `js/tests/lint.rs`.

Syscall ABI fix for wasm64:
  - `js/src/syscall_imports.rs`: `__syscall_openat`'s `mode` arg
    is declared `intptr_t` (i64 on wasm64) in Emscripten's standalone
    libc, but our Rust override used `c_int` (i32). Changed to
    `isize` — i32 on wasm32, i64 on wasm64, matches both. No
    behavior change on wasm32.

What it requires (out-of-repo)
------------------------------

The build script fails fast with a clear message if these aren't set
up:

1. **Nightly Rust toolchain** with rust-src component. `-Z build-std`
   compiles libstd for the custom JSON target.

2. **A one-line patch in nightly's rustlib** to add a
   wasm64-emscripten arm in `library/unwind/src/libunwind.rs`. Upstream
   PR pending; reference commit at:
     https://github.com/nobodywho-ooo/rust/commit/1958efbecba5106dfb95d9c93412fd132eab5b76
   The build script detects the missing arm and instructs the
   contributor on what to add.

3. **Emscripten fork at `/Users/user/git/emscripten-wbg`** (same as
   wasm32 build). First run of the script invokes `embuilder build
   SYSTEM --wasm64 -f` to populate the wasm64 sysroot with native-EH
   libcxx variants. Cached after; ~6 min one-time.

4. **`llama-cpp-rs` fork patches** for wasm64 target recognition,
   `-sMEMORY64=1`, `-fwasm-exceptions`, `EMSCRIPTEN_SYSTEM_PROCESSOR`,
   bindgen target-triple parameterization. These live in our
   `nobodywho-ooo/llama-cpp-rs` checkout (the wasm-emscripten
   branch) — not yet pushed to a tag, applied locally via cargo's git
   cache for now. Productizing means pushing to a wasm64 branch on
   the fork and pinning via `core/Cargo.toml`.

Verified
--------

- Native: `cargo check --workspace` clean (only pre-existing
  uniffi/flutter deprecation warnings, no regressions from the cfg
  refactor).

- wasm32 build path unaffected: existing
  `build-pkg-emscripten.sh` flow unchanged.

- wasm64 build via the new script, end-to-end:
    Qwen3-0.6B-Q4_K_M (484 MB):
      Chat.create 501 ms, 9 streamed chunks, ttft 1581 ms, total
      2146 ms. PASS.

    Gemma 3 4B Q4_K_M (2.5 GB) + mmproj-gemma-3-f16 (851 MB) — the
    original motivating use case:
      Chat.create 21.5 s, 10 streamed chunks, ttft 32.5 s, total
      48.1 s. PASS.

    Both via `modelPath` (no main-thread Buffer of the model bytes;
    chunked-stream into MEMFS).

Performance characteristics
---------------------------

Wasm64 inference is roughly 2-3× slower than wasm32 on the same model
(Qwen3-0.6B ttft: 1581 ms wasm64 vs ~600-700 ms wasm32). That's
expected: 8-byte pointers double the memory traffic in pointer-heavy
code paths. The wasm64 build is the right choice for "models too big
to fit in wasm32 memory", not as a general replacement. Both builds
should ship; app code picks at install time or via runtime feature
detection.

Productization status
---------------------

This is the spike output, productized to the level of "the build
script lives in the repo and works locally on a properly-set-up
machine." Two remaining gaps before this is consumer-ready:

1. **Upstream the Rust libunwind patch** so contributors don't need
   the manual rustlib edit. PR shaped + ready at
   https://github.com/nobodywho-ooo/rust/tree/wasm64-emscripten-unwind-arm

2. **Push the llama-cpp-rs build.rs changes** to a branch on
   `nobodywho-ooo/llama-cpp-rs` so the pin is reproducible without
   per-developer cargo cache patches.

Until then, this is "spike-quality build pipeline" — the bones are
real, the toolchain dependency story is rough.

+++ nobodywho/js/src/syscall_imports.rs
@@ -24,7 +24,7 @@
 //! `pkg-bundler/pre.js` at module init and lives for the lifetime of the
 //! wasm instance.
 
-#![cfg(target_arch = "wasm32")]
+#![cfg(target_family = "wasm")]
 
 use std::ffi::CStr;
 use std::os::raw::{c_char, c_int};
@@ -88,12 +88,16 @@ fn fs_call(fs: &JsValue, method: &str, args: &[JsValue]) -> Result<JsValue, c_in
 /// `path` must be a NUL-terminated C string valid for the duration of the
 /// call. Callers are libc's openat path which always passes a valid C
 /// string from the user's `fopen` argument.
+// `mode` is declared as `intptr_t` in Emscripten's wasm64 standalone
+// libc (i64 on wasm64, i32 on wasm32). Using `isize` makes the Rust
+// signature match the linker's expectation on both pointer widths.
+// We narrow back to c_int when calling FS.open.
 #[no_mangle]
 pub unsafe extern "C" fn __syscall_openat(
     _dirfd: c_int,
     path: *const c_char,
     flags: c_int,
-    mode: c_int,
+    mode: isize,
 ) -> c_int {
     let fs = match fs_namespace() {
         Ok(fs) => fs,
